pyproject.toml
[tool.poetry]
authors = ["Laxman Desai <lolopaw24@gmail.com>"]
description = ""
license = "MIT"
name = "data-entry-ai"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
gradio = "^5.4.0"
langchain = "^0.3.4"
langchain-community = "^0.3.3"
langchain-core = "^0.3.13"
langchain-openai = "^0.2.4"
numpy = "^1.26.4"
pandas = "^2.2.3"
pillow = "^10.4.0"
pymupdf = "^1.24.13"
pytesseract = "^0.3.13"
python = "~3.12"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.2.1"
streamlit = "^1.39.0"
torch = "^2.5.1"
tqdm = "^4.66.6"
transformers = "^4.46.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[tool.isort]
line_length = 88
profile = "black"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
